Refactor the database seed script
