+++ prisma/seed/index.ts
@@ -1,0 +1,2 @@
+export { default } from './run';
+export * from './run';
